refactor Home page with Messages
