+++ src/components/Messages/Messages.js
@@ -1,0 +1,154 @@
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { compose } from 'recompose';
+
+import { AuthUserContext } from '../Session';
+import { withFirebase } from '../Firebase';
+import MessageList from './MessageList';
+
+class Messages extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      text: '',
+      loading: false,
+    };
+  }
+
+  componentDidMount() {
+    if (!this.props.messages.length) {
+      this.setState({ loading: true });
+    }
+
+    this.onListenForMessages();
+  }
+
+  componentDidUpdate(props) {
+    if (props.limit !== this.props.limit) {
+      this.onListenForMessages();
+    }
+  }
+
+  onListenForMessages = () => {
+    this.props.firebase
+      .messages()
+      .orderByChild('createdAt')
+      .limitToLast(this.props.limit)
+      .on('value', snapshot => {
+        this.props.onSetMessages(snapshot.val());
+
+        this.setState({ loading: false });
+      });
+  };
+
+  componentWillUnmount() {
+    this.props.firebase.messages().off();
+  }
+
+  onChangeText = event => {
+    this.setState({ text: event.target.value });
+  };
+
+  onCreateMessage = (event, authUser) => {
+    this.props.firebase.messages().push({
+      text: this.state.text,
+      userId: authUser.uid,
+      createdAt: this.props.firebase.serverValue.TIMESTAMP,
+    });
+
+    this.setState({ text: '' });
+
+    event.preventDefault();
+  };
+
+  onEditMessage = (message, text) => {
+    this.props.firebase.message(message.uid).set({
+      ...message,
+      text,
+      editedAt: this.props.firebase.serverValue.TIMESTAMP,
+    });
+  };
+
+  onRemoveMessage = uid => {
+    this.props.firebase.message(uid).remove();
+  };
+
+  onNextPage = () => {
+    this.props.onSetMessagesLimit(this.props.limit + 5);
+  };
+
+  render() {
+    const { users, messages } = this.props;
+    const { text, loading } = this.state;
+
+    return (
+      <AuthUserContext.Consumer>
+        {authUser => (
+          <div>
+            {!loading && messages && (
+              <button type="button" onClick={this.onNextPage}>
+                More
+              </button>
+            )}
+
+            {loading && <div>Loading ...</div>}
+
+            {messages && (
+              <MessageList
+                messages={messages.map(message => ({
+                  ...message,
+                  user: users
+                    ? users[message.userId]
+                    : { userId: message.userId },
+                }))}
+                onEditMessage={this.onEditMessage}
+                onRemoveMessage={this.onRemoveMessage}
+              />
+            )}
+
+            {!messages && <div>There are no messages ...</div>}
+
+            <form
+              onSubmit={event =>
+                this.onCreateMessage(event, authUser)
+              }
+            >
+              <input
+                type="text"
+                value={text}
+                onChange={this.onChangeText}
+              />
+              <button type="submit">Send</button>
+            </form>
+          </div>
+        )}
+      </AuthUserContext.Consumer>
+    );
+  }
+}
+
+const mapStateToProps = state => ({
+  messages: Object.keys(state.messageState.messages || {}).map(
+    key => ({
+      ...state.messageState.messages[key],
+      uid: key,
+    }),
+  ),
+  limit: state.messageState.limit,
+});
+
+const mapDispatchToProps = dispatch => ({
+  onSetMessages: messages =>
+    dispatch({ type: 'MESSAGES_SET', messages }),
+  onSetMessagesLimit: limit =>
+    dispatch({ type: 'MESSAGES_LIMIT_SET', limit }),
+});
+
+export default compose(
+  withFirebase,
+  connect(
+    mapStateToProps,
+    mapDispatchToProps,
+  ),
+)(Messages);
