+++ src/components/Messages/index.js
@@ -1,0 +1,3 @@
+import Messages from './Messages';
+
+export default Messages;
